Rename `preprocess` and `postprocess` file
