+++ src/language-js/parser-typescript.js
@@ -3,7 +3,7 @@
 const createError = require("../common/parser-create-error");
 const { hasPragma } = require("./pragma");
 const { locStart, locEnd } = require("./loc");
-const postprocess = require("./postprocess");
+const postprocess = require("./parse-postprocess");
 
 function parse(text, parsers, opts) {
   const jsx = isProbablyJsx(text);
